+++ lib/xgen/generators/elixir/nerves.ex
@@ -32,7 +32,7 @@ defmodule XGen.Generators.Elixir.Nerves do
   ]
 
   overrides %{
-    initial_version: "0.0.1-dev",
+    initial_version: "0.0.1",
     module_path: Macro.underscore(@module),
     cookie: generate_cookie(),
     sup?: @sup? || @ssh?