+++ lib/xgen/options/elixir/std.ex
@@ -4,9 +4,34 @@ defmodule XGen.Options.Elixir.Std do
   use XGen.Option, collection: true
 
   defoption Release do
+    defoption ConfigProvider do
+      key :config_provider
+      type :choice
+      default :toml
+      options choices: config_providers()
+      name "Distillery “Config Providers”"
+      prompt "Which Config Provider do you want to use?"
+
+      documentation """
+      You may need to configure your application after it has been compiled,
+      when deploying your OTP release. Distillery comes with a configuration
+      framework called “Config Providers”. xgen can generate a template for you
+      using different config providers.
+      """
+
+      defp config_providers do
+        [
+          toml: "Toml.Provider",
+          mix: "Mix.Releases.Config.Providers.Elixir",
+          nil: "Do not use a Config Provider"
+        ]
+      end
+    end
+
     key :release?
     type :yesno
     default :no
+    options if_yes: [ConfigProvider]
     name "Release"
     prompt "Generate a release configuration?"